pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "webclipper"
version = "0.1.2"
authors = [
  { name="Hugo Martiniano", email="hugomartiniano@gmail.com" },
]
description = "A tool to fetch the main content of a webpage and convert it to Markdown or plain text."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "beautifulsoup4",
    "html2text",
    "readability-lxml",
]

[project.urls]
Homepage = "https://github.com/hmartiniano/webclipper" # Optional

# This section is crucial for creating the command-line app.
# It tells the installer to create a script named 'webclipper'
# that calls the 'main' function inside your 'main.py' module.
[project.scripts]
webclipper = "webclipper.main:main"

